requirements.txt
python-dotenv==0.19.2
pytube==11.0.1
pytube3==9.6.4
pyudev==0.22.0
six==1.16.0
typing_extensions==4.0.1
